+++ hutool-core/src/main/java/cn/hutool/core/net/url/UrlBuilder.java
@@ -83,6 +83,17 @@ public final class UrlBuilder implements Serializable {
 	}
 
 	/**
+	 * 使用URL字符串构建UrlBuilder，当传入的URL没有协议时，按照http协议对待，编码默认使用UTF-8
+	 *
+	 * @param httpUrl URL字符串
+	 * @return UrlBuilder
+	 * @since 5.6.3
+	 */
+	public static UrlBuilder ofHttp(String httpUrl) {
+		return ofHttp(httpUrl, CharsetUtil.CHARSET_UTF_8);
+	}
+
+	/**
 	 * 使用URL字符串构建UrlBuilder，当传入的URL没有协议时，按照http协议对待。
 	 *
 	 * @param httpUrl URL字符串
@@ -97,6 +108,16 @@ public final class UrlBuilder implements Serializable {
 			httpUrl = "http://" + httpUrl.trim();
 		}
 		return of(httpUrl, charset);
+	}
+
+	/**
+	 * 使用URL字符串构建UrlBuilder，默认使用UTF-8编码
+	 *
+	 * @param url     URL字符串
+	 * @return UrlBuilder
+	 */
+	public static UrlBuilder of(String url) {
+		return of(url, CharsetUtil.CHARSET_UTF_8);
 	}
 
 	/**